Rename interactor io header
